requirements.txt
#aioredis==2.0.1
#aiosqlite==0.17.0
#alembic==1.11.1
aio-pika==9.5.5
asyncpg==0.29.0
aerich==0.7.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
bcrypt==4.2.0
black==23.3.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.2
fastapi-limiter==0.1.6
flake8==6.0.0
h11==0.14.0
httpcore==1.*
httplib2==0.22.0
httptools==0.6.1
httpx>=0.28.1
idna==3.8
importlib_resources==6.4.5
iniconfig==2.0.0
iso8601==1.1.0
isort==5.12.0
limits==3.13.0
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
msgpack==1.0.8
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.2
pluggy==1.5.0
pony==0.7.19
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
#pydantic>=2.10.3,<3.0.0
#pydantic-settings<=2.0.0
pydantic_core>=2.27.0
pyflakes==3.0.1
PyJWT==2.9.0
PyMySQL==1.0.3
pyparsing==3.1.4
pypika-tortoise==0.1.6
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-env==1.0.1
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
rsa==4.9
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.38.5
tomli==2.0.1
tortoise-orm==0.21.5
uvicorn==0.34.1
clerk-backend-api==2.0.2
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-instrumentation-logging==0.56b0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-instrumentation-httpx==0.56b0
opentelemetry-instrumentation-redis==0.56b0
opentelemetry-exporter-otlp-proto-http==1.35.0
prometheus-fastapi-instrumentator==7.1.0
